==> 1-ZipCode_Master.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler,OneHotEncoder
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import io

# Load the zipcode dataset
# Load zipcode dataset and ensure that zipcodes are read as strings
file_path = "data/USZip_info.csv"
zip_df = pd.read_csv(file_path,dtype={"zip":"string"})
zip_df.head(10)

zipcode_df = zip_df[['zip', 'lat', 'lng', 'city', 'state_id', 'state_name', 'zcta', 'population', 'density', 'county_fips']]
zipcode_df.head(10)

# Rename column titles
zipcode_df = zipcode_df.rename(columns={'zip': 'ZipCode'})
zipcode_df.head(5)

zipcode_df.dtypes

zipcode_df.to_csv('data/zipcode_demo.csv',index=False)
