==> 2-Customer_RFM_Enhancement.ipynb <==
import pandas as pd
from datetime import timedelta
import numpy as np

# Read datasets in
> Ensure that zipcdes are read as strings and not numeric

# Load the customer dataset
file_path = "data/Customer_sales.csv"
df_customer_sales = pd.read_csv(file_path,dtype={"BillPostalCode":"string"} )
df_customer_sales
df_customer_sales["CustomerID"].nunique()

# Group by order and customer - total sales
Orders_df = df_customer_sales.groupby(['ChannelOrderID','OrderDate','CustomerID']).agg({'Total': lambda x:x.sum()}).reset_index()
Orders_df

# Convert orderdate to datetime type so calculations can be done
Orders_df['OrderDate'] = Orders_df['OrderDate'].astype('datetime64[ns]')

Orders_df.dtypes

# Set variables for NOW and and period
NOW = Orders_df['OrderDate'].max() + timedelta(days=1)
NOW

# Data is for a two year period
period = 730

# Add column for Days since order to aid with calcuation
Orders_df['DaysSinceOrder'] = Orders_df['OrderDate'].apply(lambda  x: (NOW -x).days)

aggr = {
    'DaysSinceOrder' : lambda x : x.min(),
    'OrderDate' : lambda x: len([d for d in x if d>=NOW - timedelta(days=period)])
}

rfm_df = Orders_df.groupby('CustomerID').agg(aggr).reset_index()
rfm_df.rename(columns={'DaysSinceOrder' : 'Recency', 'OrderDate' : 'Frequency'}, inplace=True)
rfm_df

# rfm_df['Monetary'] = rfm_df['CustomerID'].apply(lambda x: Orders_df[('CustomerID'] == x) & Orders_df
# ['OrderDate'] >= NOW - timedelta(days=period)['Total'].sum())




rfm_df['Monetary'] = rfm_df['CustomerID'].apply(lambda x: Orders_df[(Orders_df['CustomerID'] == x) &
                                                           (Orders_df['OrderDate'] >= NOW - timedelta(days=period))]
                                                           ['Total'].sum())

rfm_df

quintiles = rfm_df[['Recency', 'Frequency', 'Monetary']].quantile([.2, .4, .6, .8]).to_dict()
quintiles

def r_score(x):
    if x <= quintiles['Recency'][.2]:
        return 5
    elif x <= quintiles['Recency'][.4]:
        return 4
    elif x <= quintiles['Recency'][.6]:
        return 3
    elif x <= quintiles['Recency'][.8]:
        return 2
    else:
        return 1

def fm_score(x, c):
    if x <= quintiles[c][.2]:
        return 1
    elif x <= quintiles[c][.4]:
        return 2
    elif x <= quintiles[c][.6]:
        return 3
    elif x <= quintiles[c][.8]:
        return 4
    else:
        return 5

rfm_df['R'] = rfm_df['Recency'].apply(lambda x: r_score(x))
rfm_df['F'] = rfm_df['Frequency'].apply(lambda x: fm_score(x, 'Frequency'))
rfm_df['M'] = rfm_df['Monetary'].apply(lambda x: fm_score(x, 'Monetary'))

rfm_df

rfm_df['RFM Score'] = rfm_df['R'].map(str) + rfm_df['F'].map(str) + rfm_df['M'].map(str)
rfm_df.head()

segt_map = {
    r'[1-2][1-2]': 'hibernating',
    r'[1-2][3-4]': 'at risk',
    r'[1-2]5': 'can\'t loose',
    r'3[1-2]': 'about to sleep',
    r'33': 'need attention',
    r'[3-4][4-5]': 'loyal customers',
    r'41': 'promising',
    r'51': 'new customers',
    r'[4-5][2-3]': 'potential loyalists',
    r'5[4-5]': 'champions'
}

rfm_df['Segment'] = rfm_df['R'].map(str) + rfm_df['F'].map(str)
rfm_df['Segment'] = rfm_df['Segment'].replace(segt_map, regex=True)
rfm_df

rfm_df.to_csv('data/Customer_rfm.csv', index=False)
