==> 3-Customer_Sales_Pivot.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler,OneHotEncoder
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import io

## Load the customer sales dataset

# Load the customer dataset
file_path = "data/Customer_sales.csv"
df_customer_sales = pd.read_csv(file_path,dtype={"BillPostalCode":"string"} )
df_customer_sales

## Pivot and output brand totals by Customer

df_customer_sales['Brand'].unique()


# Limit to the columns want to pivot
df_customer_sales_brand_data = df_customer_sales[['CustomerID', 'Brand', 'Total']]

# Group by customer and pivot column
df_group_by_brand = df_customer_sales_brand_data.groupby(['CustomerID','Brand']).sum()

# Unstack coluumns
df_group_by_brand_pivot = df_group_by_brand.unstack()

# Fill in nulls with zeros
df_group_by_brand_pivot.fillna(0, inplace=True)
df_group_by_brand_pivot.columns = df_group_by_brand_pivot.columns.to_flat_index()

column_names = {('Total', 'Arctic Cat') :'ArticCat_Tot', ('Total', 'Bobcat'): "Bobcat_Tot",
                ('Total', 'Can-Am'): 'Can-Am_Tot', ('Total', 'Honda') : 'Honda_Tot',
                ('Total', 'John Deere') : 'John-Deere_Tot', ('Total', 'Kawasaki') : 'Kawasaki_Tot',
                ('Total', 'Kubota') : 'Kubota_Tot', ('Total', 'Other-Brand') : 'Other-Brand_Tot',  ('Total', 'Polaris') : 'Polaris_Tot',
                ('Total', 'Suzuki') : 'Suzuki_Tot', ('Total', 'Yamaha') : 'Yamaha_Tot'}

df_group_by_brand_pivot = df_group_by_brand_pivot.rename(columns=column_names)

df_group_by_brand_pivot

# Output to CSV file
df_group_by_brand_pivot.to_csv('data/Customer_brands.csv')

## Pivot and output LOB totals by Customer

# Limit to the columns want to pivot
df_customer_sales_lob_data = df_customer_sales[['CustomerID', 'LOB', 'Total']]

# Group by customer and pivot column
df_group_by_lob = df_customer_sales_lob_data.groupby(['CustomerID','LOB']).sum()

# Unstack coluumns
df_group_by_lob_pivot = df_group_by_lob.unstack()

# Fill in nulls with zeros
df_group_by_lob_pivot.fillna(0, inplace=True)


df_group_by_lob_pivot.columns = df_group_by_lob_pivot.columns.to_flat_index()

df_group_by_lob_pivot

column_names = {('Total', 'Misc-LOB') :'LOB_Misc_Tot', ('Total', 'NOS-Rebuilt-Parts'): "LOB_NOS-Rebuilt-Part_Tot",
               ('Total', 'Rebuilt-Engines'): 'LOB_Rebuilt-Engines-Tot', ('Total', 'Salvage') : 'LOB_Salvage_Tot'}

df_group_by_lob_pivot = df_group_by_lob_pivot.rename(columns=column_names)

df_group_by_lob_pivot

### Add logical LOB columns

# Add Logical LOB Business
df_group_by_lob_pivot['LOB_Misc'] = df_group_by_lob_pivot['LOB_Misc_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_lob_pivot['LOB_NOS-Rebuilt-Part'] = df_group_by_lob_pivot['LOB_NOS-Rebuilt-Part_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_lob_pivot['LOB_Salvage'] = df_group_by_lob_pivot['LOB_Salvage_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_lob_pivot['LOB_Rebuilt-Engines'] = df_group_by_lob_pivot['LOB_Rebuilt-Engines-Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_lob_pivot

# Output to CSV file
df_group_by_lob_pivot.to_csv('data/Customer_lob.csv')

## Pivot and output Market totals by Customer

# Limit to the columns want to pivot
df_customer_sales_market_data = df_customer_sales[['CustomerID', 'Market', 'Total']]

# Group by customer and pivot column
df_group_by_market = df_customer_sales_market_data.groupby(['CustomerID','Market']).sum()

# Unstack coluumns
df_group_by_market_pivot = df_group_by_market.unstack()

# Fill in nulls with zeros
df_group_by_market_pivot.fillna(0, inplace=True)


df_group_by_market_pivot.columns = df_group_by_market_pivot.columns.to_flat_index()

df_group_by_market_pivot.columns

column_names = {('Total', 'Counter') :'Mkt-Counter_Tot', ('Total', 'Other-Market'): "Mkt-Other_Tot",
                ('Total', 'Website'): 'Mkt-Website_Tot', ('Total', 'eBay') : 'Mkt-eBay_Tot'}

df_group_by_market_pivot = df_group_by_market_pivot.rename(columns=column_names)

df_group_by_market_pivot

# Add Logical Market Columns
df_group_by_market_pivot['Buyer_eBay'] = df_group_by_market_pivot['Mkt-eBay_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_market_pivot['Buyer_website'] = df_group_by_market_pivot['Mkt-Website_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_market_pivot['Buyer_counter'] = df_group_by_market_pivot['Mkt-Counter_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_market_pivot['Buyer_other'] = df_group_by_market_pivot['Mkt-Other_Tot'].apply(lambda x: 1 if x >0 else 0)

df_group_by_market_pivot

# Output to CSV file
df_group_by_market_pivot.to_csv('data/Customer_market.csv')
