==> 4-Customer_Data_Merge.ipynb <==
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler,OneHotEncoder
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import io

## Load Brand CSV File and Merge to Customer Master

# Load the customer dataset
file_path = "data/Customers.csv"
df_customers = pd.read_csv(file_path,dtype={"ZipCode":"string"} )
df_customers

# Load the customer brand dataset
file_path = "data/Customer_brands.csv"
df_customer_brands = pd.read_csv(file_path)
df_customer_brands

df_customer_data  = pd.merge(df_customers, df_customer_brands, left_on='CustomerID', right_on='CustomerID')
df_customer_data

## Load LOB CSV File and Merge to Customer Master

# Load the customer lob dataset
file_path = "data/Customer_lob.csv"
df_customer_lobs = pd.read_csv(file_path )
df_customer_lobs

df_customer_data  = pd.merge(df_customer_data, df_customer_lobs, left_on='CustomerID', right_on='CustomerID')
df_customer_data

## Load Market CSV File and Merge to Customer Master

# Load the customer lob dataset
file_path = "data/Customer_market.csv"
df_customer_market = pd.read_csv(file_path )
df_customer_market

df_customer_data  = pd.merge(df_customer_data, df_customer_market, left_on='CustomerID', right_on='CustomerID')
df_customer_data

## Load RFM CSV File and Merge to Customer Master

# Load the customer lob dataset
file_path = "data/Customer_rfm.csv"
df_customer_rfm = pd.read_csv(file_path )
df_customer_rfm

df_customer_data  = pd.merge(df_customer_data, df_customer_rfm, left_on='CustomerID', right_on='CustomerID')
df_customer_data

## Merge Demographics into the Customer Merge

# Load the zipcode demographics dataset
file_path = "data/zipcode_demo.csv"
df_zipcodes = pd.read_csv(file_path,dtype={"ZipCode":"string"})
df_zipcodes

df_customer_data = pd.merge(df_customer_data, df_zipcodes, left_on='ZipCode', right_on='ZipCode')
df_customer_data

## Write out Merge Customer Data File
> Need to merge additional demographics for this file. This file contains a single row with roll-up data into columns per customer

df_customer_data.to_csv('data/Customer_data_merged.csv',index=False)
